# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Full Stack Svelte Template API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              }
            },
            "required": [
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "GoalStatus": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "forecast",
              "past",
              "beyond"
            ]
          },
          "expectedCents": {
            "type": "integer",
            "nullable": true
          },
          "worstCents": {
            "type": "integer",
            "nullable": true
          },
          "bestCents": {
            "type": "integer",
            "nullable": true
          },
          "monthlyNeededCents": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "state",
          "expectedCents",
          "worstCents",
          "bestCents",
          "monthlyNeededCents"
        ]
      },
      "Goal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "amountCents": {
            "type": "integer"
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "example": "2026-10-01"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "amountCents",
          "targetDate",
          "createdAt",
          "updatedAt",
          "status"
        ]
      },
      "GoalList": {
        "type": "object",
        "properties": {
          "goals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Goal"
            }
          },
          "today": {
            "type": "string",
            "format": "date",
            "example": "2026-10-01"
          }
        },
        "required": [
          "goals",
          "today"
        ]
      },
      "CreateGoalBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Trip home"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 280,
            "example": "Flights for December"
          },
          "amountCents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000000,
            "example": 150000,
            "description": "integer cents, more than 0"
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "example": "2026-10-01"
          }
        },
        "required": [
          "name",
          "amountCents",
          "targetDate"
        ]
      },
      "UpdateGoalBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Trip home"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 280,
            "example": "Flights for December"
          },
          "amountCents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000000,
            "example": 150000,
            "description": "integer cents, more than 0"
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "example": "2026-10-01"
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/api/goals": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "List goals with forecast status, closest date first",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Goals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalList"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Create a goal",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/goals/{id}": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "Get a goal",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Goal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "goals"
        ],
        "summary": "Update a goal",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "goals"
        ],
        "summary": "Delete a goal",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  }
}